Fixing wiserf classifier

+++ EmailAnalysis/src/main/java/EmailAnalysis/Feature.java
@@ -3,8 +3,9 @@ package EmailAnalysis;
 public interface Feature {
     /**
      * Returns the value of this feature given an email.
+     *
      * @param email
      * @return The value of the feature. No scale is specified.
      */
-    public float getValue(CleanedEmail email);
+    public double getValue(CleanedEmail email);
 }
